requirements.txt
tqdm
sklearn
numpy
matplotlib
pillow
